requirements.txt
authentik-client==2025.2.2.post1742395435
fastapi==0.115.12
outline-python-api==0.4.0a0
python-dotenv==1.1.0
uvicorn==0.34.0
httpx==0.27.2